Projects/LLMProteinOptimizer/requirements.txt
# Core ML/DL
torch==2.1.2
torchvision==0.16.2
transformers==4.47.1
accelerate==0.33.0
numpy==1.26.4
pandas==2.2.3
scipy==1.13.1
scikit-learn==1.2.2
xgboost==2.1.3
omegaconf
Levenshtein

# Optimization & Modeling
botorch==0.10.0
gpytorch==1.11
pymoo==0.6.1.3
cma==3.2.2

# Visualization
matplotlib==3.9.4
seaborn==0.13.2
plotly==5.24.1

# Utilities
tqdm==4.67.1
joblib==1.4.2
pyyaml==6.0.2
progressbar

# Optional but recommended
biopython==1.84  # For protein sequence handling
h5py==3.12.1    # For data storage